+++ tests/tst_main.cpp
@@ -12,6 +12,7 @@ int runSessionTests(int argc, char **argv);
 int runControllerTests(int argc, char **argv);
 int runCommandTests(int argc, char **argv);
 int runIgnoreTests(int argc, char **argv);
+int runMuteTests(int argc, char **argv);
 int runTypingTests(int argc, char **argv);
 int runProfileTests(int argc, char **argv);
 int runConnectionTests(int argc, char **argv);
@@ -42,6 +43,7 @@ int main(int argc, char **argv)
     const int controllerStatus = runControllerTests(argc, argv);
     const int commandStatus = runCommandTests(argc, argv);
     const int ignoreStatus = runIgnoreTests(argc, argv);
+    const int muteStatus = runMuteTests(argc, argv);
     const int typingStatus = runTypingTests(argc, argv);
     const int profileStatus = runProfileTests(argc, argv);
     const int connectionStatus = runConnectionTests(argc, argv);
@@ -64,6 +66,7 @@ int main(int argc, char **argv)
         controllerStatus,
         commandStatus,
         ignoreStatus,
+        muteStatus,
         typingStatus,
         profileStatus,
         connectionStatus,